Move constants Categroy/Management to helpers
