+++ cmd/nerdctl/volume.go
@@ -18,11 +18,13 @@ package main
 
 import (
 	"github.com/spf13/cobra"
+
+	"github.com/containerd/nerdctl/v2/cmd/nerdctl/helpers"
 )
 
 func newVolumeCommand() *cobra.Command {
 	volumeCommand := &cobra.Command{
-		Annotations:   map[string]string{Category: Management},
+		Annotations:   map[string]string{helpers.Category: helpers.Management},
 		Use:           "volume",
 		Short:         "Manage volumes",
 		RunE:          unknownSubcommandAction,
